Fix correctness bugs + tighten domain immutability
Critical:
- GreedyKnapsack: ProcessWeldedOrders now hosts partial tail pieces in
  existing non-welded plans' leftover instead of always burning a fresh
  bar (FindHostPlanForWeld).
- GreedyKnapsack: FindTopKCutsSparse DP dedup switched from List to
  Dictionary so the smallest-cut-count candidate is always kept per key.
- SolverUtils: RedistributeCuts now guards welded cuts on both source
  and destination, matching LocalSearchOptimize.
- ArcFlow: DecomposeFlowToPatterns has a safety counter to prevent
  degenerate flows from infinite-looping.
- PatternBuilder: BuildTree now handles single-rect non-corner cases by
  peeling waste off each side recursively (was returning null).
- 2D solvers: TimeLimitMs is now an absolute deadline from session
  start (was double-counting bootstrap time).
- TotalCost widened to long across SolverResult / ComparisonResult /
  ExportService to prevent overflow on large inputs.

Quality:
- Order, RebarStock now fully immutable (constructor-validated, get-only).
- Cut and CuttingPlan.StockLength / .Cuts switched to init-only.
- 2D solvers share SolverUtils2D.ValidateInputs instead of duplicating
  null/empty guards.
- EstimateFutureWasteFFDFromDict renamed to MFFD to match implementation.
- GenerateSolutionHybrid renamed to ApplyGreedyResidual with docstring.

Performance:
- GreedyKnapsack candidate scoring no longer copies a Dictionary per
  candidate: reusable workingQty buffer + appliedCuts log.

Tests: 521/521 (including +3 regression tests). LargeScale benchmark
verified passes in ~7s.

+++ CuttingStock.Benchmarks/QualityBenchmarks.cs
@@ -22,7 +22,7 @@ namespace CuttingStock.Benchmarks
         private SolverOptions _defaultParams = null!;
 
         // 결과 저장용
-        public int LastTotalCost { get; private set; }
+        public long LastTotalCost { get; private set; }
         public int LastWasteLength { get; private set; }
         public int LastStockUsed { get; private set; }
         public double LastMaterialEfficiency { get; private set; }
